mescc: Mes C Library: Add x86_64 libc-mini support.
* lib/linux/x86-mes-gcc/mini.c: Rename from lib/linux/mini-gcc.c.
* lib/linux/x86-mes/mini.c: Rename from lib/linux/mini-mes.c.
* lib/linux/x86_64-mes-gcc/crt1.c: Rename from lib/linux/x86_64/crt1.c.
* lib/linux/x86_64-mes-gcc/mini.c: New file.
* build-aux/build-cc.sh: Build libc-mini, scaffold/main, scaffold/hello.
* build-aux/cc64.sh: New file.

+++ scaffold/mini-mes.c
@@ -38,7 +38,7 @@ int JAM_SIZE = 20000;
 int GC_SAFETY = 2000;
 
 char *g_arena = 0;
-typedef int SCM;
+typedef long SCM;
 
 int g_debug = 0;
 int g_free = 0;
